+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = ["Fábio Madeira <fmadeira@ebi.ac.uk>"]
 readme = "README.rst"
 include = ["CHANGELOG.rst"]
 license = "Apache-2.0"
-packages = [{include = "taxonresolver"}]
+packages = [{include = "taxonomyresolver"}]
 classifiers = [
         "Development Status :: 4 - Beta",
         "Intended Audience :: Developers",
