+++ taxonomyresolver/__init__.py
@@ -13,4 +13,4 @@ __email__ = "fmadeira@ebi.ac.uk"
 __version__ = "0.3.2"
 __contributors__ = ["Fábio Madeira"]
 
-from taxonresolver.tree import TaxonResolver
+from taxonomyresolver.tree import TaxonResolver
